Extract UsageFailure into usage_failure.dart
