+++ lib/src/sdk/errors.dart
@@ -1,12 +1,5 @@
 import 'failure.dart';
 
-class UsageFailure extends Failure {
-  UsageFailure({
-    required String usage,
-    Map<String, String>? context,
-  }) : super(message: usage, context: context);
-}
-
 class AlreadyAppliedFailure extends Failure {
   AlreadyAppliedFailure({
     required String action,
