+++ lib/src/sdk/sdk.dart
@@ -17,3 +17,4 @@ export 'trello_authentication.dart';
 export 'trello_client.dart';
 export 'trello_models.dart';
 export 'trello_object.dart';
+export 'usage_failure.dart';
