+++ lib/src/sdk/usage_failure.dart
@@ -1,0 +1,8 @@
+import 'failure.dart';
+
+class UsageFailure extends Failure {
+  UsageFailure({
+    required String usage,
+    Map<String, String>? context,
+  }) : super(message: usage, context: context);
+}
